+++ gqlgen.yml
@@ -10,7 +10,7 @@ exec:
 model:
   filename: models_gen.go
 resolver:
-  filename: resolver.go
+  filename: resolver/resolver.go
   type: Resolver
 models:
   ID: # The GraphQL type ID is backed by
